Create 1-introduction.md
Hi, here's some textual changes! Suggest to add link to the images as the text suggests you can click them. I understand we mean in the platform, but not sure this is understood by every new hub member: for instance 'This is done by clicking on the cogwheel icon on the right:'